requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
segment-anything-2>=0.1.0
diffusers>=0.25.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
click>=8.1.0
webp>=0.3.0
pytest>=8.0.0
